ser-libs/capnproto/demo-01/pyproject.toml
[tool.poetry]
name = "demo-01"
version = "0.1.0"
description = ""
authors = ["Felipe Bidu Rodrigues <felipe@felipevr.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pycapnp = "^1.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = {version = "^21.8b0", allow-prereleases = true}
pyprof2calltree = "^1.4.5"
memory-profiler = "^0.58.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
